+++ packages/dash-table/src/dash-table/components/HeaderFactory.tsx
@@ -36,6 +36,7 @@ export default class HeaderFactory {
         const props = this.props;
 
         const {
+            column_selectable,
             columns,
             data,
             hidden_columns,
@@ -44,6 +45,7 @@ export default class HeaderFactory {
             page_action,
             row_deletable,
             row_selectable,
+            selected_columns,
             setFilter,
             setProps,
             sort_action,
@@ -92,10 +94,13 @@ export default class HeaderFactory {
 
         const contents = this.headerContent(
             visibleColumns,
+            columns,
             hidden_columns,
             data,
             labelsAndIndices,
             map,
+            column_selectable,
+            selected_columns,
             sort_action,
             sort_mode,
             sort_by,