+++ packages/dash-table/src/dash-table/derived/header/content.tsx
@@ -8,13 +8,14 @@ import singleUpdate from 'core/sorting/single';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
     ColumnId,
-    Data,
-    SortMode,
     Columns,
+    Data,
     IColumn,
+    Selection,
+    SetFilter,
     SetProps,
-    TableAction,
-    SetFilter
+    SortMode,
+    TableAction
 } from 'dash-table/components/Table/props';
 import getColumnFlag from 'dash-table/derived/header/columnFlag';
 import * as actions from 'dash-table/utils/actions';
@@ -29,6 +30,7 @@ const doAction = (
         mergeDuplicateHeaders: boolean,
         data: Data
     ) => any,
+    selected_columns: string[],
     column: IColumn,
     columns: Columns,
     columnRowIndex: any,
@@ -38,7 +40,19 @@ const doAction = (
     map: Map<string, SingleColumnSyntaxTree>,
     data: Data
 ) => () => {
-    setProps(action(column, columns, columnRowIndex, mergeDuplicateHeaders, data));
+    const props = action(column, columns, columnRowIndex, mergeDuplicateHeaders, data);
+
+    const affectedColumIds = actions.getAffectedColumns(column, columns, columnRowIndex, mergeDuplicateHeaders);
+
+    if (action === actions.deleteColumn) {
+        if (R.intersection(selected_columns, affectedColumIds).length > 0) {
+            props.selected_columns = R.reject(
+                R.contains(R.__, affectedColumIds),
+                selected_columns
+            );
+        }
+    }
+    setProps(props);
 
     const affectedColumns: Columns = [];
     R.forEach(id => {
@@ -46,7 +60,7 @@ const doAction = (
         if (affectedColumn) {
             affectedColumns.push(affectedColumn);
         }
-    }, actions.getAffectedColumns(column, columns, columnRowIndex, mergeDuplicateHeaders));
+    }, affectedColumIds);
 
     clearColumnsFilter(map, affectedColumns, setFilter);
 };
@@ -89,6 +103,32 @@ function editColumnName(column: IColumn, columns: Columns, columnRowIndex: any, 
     };
 }
 
+function selectColumn(current_selection: string[], column: IColumn, columns: Columns, columnRowIndex: any, setProps: SetProps, mergeDuplicateHeaders: boolean, clearSelection: boolean, select: boolean) {
+    // if 'single' and trying to click selected radio -> do nothing
+    if (clearSelection && !select) {
+        return () => { };
+    }
+
+    let selected_columns = actions.getAffectedColumns(column, columns, columnRowIndex, mergeDuplicateHeaders, true);
+
+    if (clearSelection) {
+        return () => setProps({ selected_columns });
+    } else if (select) {
+        // 'multi' + select -> add to selection (union)
+        return () => setProps({
+            selected_columns: R.union(current_selection, selected_columns)
+        });
+    } else {
+        // 'multi' + unselect -> invert of intersection
+        return () => setProps({
+            selected_columns: R.reject(
+                R.contains(R.__, selected_columns),
+                current_selection
+            )
+        });
+    }
+}
+
 function getSorting(columnId: ColumnId, sortBy: SortBy): SortDirection {
     const sort = R.find(s => s.column_id === columnId, sortBy);
 
@@ -108,11 +148,14 @@ function getSortingIcon(columnId: ColumnId, sortBy: SortBy) {
 }
 
 function getter(
+    visibleColumns: Columns,
     columns: Columns,
     hiddenColumns: string[] | undefined,
     data: Data,
     labelsAndIndices: R.KeyValuePair<any[], number[]>[],
     map: Map<string, SingleColumnSyntaxTree>,
+    column_selectable: Selection,
+    selected_columns: string[],
     sort_action: TableAction,
     mode: SortMode,
     sortBy: SortBy,
@@ -128,7 +171,7 @@ function getter(
 
             return R.addIndex<number, JSX.Element>(R.map)(
                 (columnIndex, index) => {
-                    const column = columns[columnIndex];
+                    const column = visibleColumns[columnIndex];
 
                     let colSpan: number;
                     if (!mergeDuplicateHeaders) {
@@ -145,60 +188,105 @@ function getter(
                     const deletable = paginationMode !== TableAction.Custom && getColumnFlag(headerRowIndex, lastRow, column.deletable);
                     const hideable = getColumnFlag(headerRowIndex, lastRow, column.hideable);
                     const renamable = getColumnFlag(headerRowIndex, lastRow, column.renamable);
+                    const selectable = getColumnFlag(headerRowIndex, lastRow, column.selectable);
 
-                    const spansAllColumns = columns.length === colSpan;
+                    const spansAllColumns = visibleColumns.length === colSpan;
+
+                    const affectedColumns = actions.getAffectedColumns(
+                        column,
+                        columns,
+                        headerRowIndex,
+                        mergeDuplicateHeaders,
+                        true
+                    );
+                    const allSelected =
+                        selectable &&
+                        (
+                            column_selectable !== 'single' ||
+                            (selected_columns.length === affectedColumns.length)
+                        ) &&
+                        R.all(
+                            c => selected_columns.indexOf(c) !== -1,
+                            affectedColumns
+                        );
 
                     return (<div>
-                        {sort_action !== TableAction.None && isLastRow ?
+                        {!column_selectable || !selectable ?
+                            null :
+                            <span
+                                className='column-header--select'
+                            >
+                                <input
+                                    checked={allSelected}
+                                    onChange={selectColumn(
+                                        selected_columns,
+                                        column,
+                                        columns,
+                                        headerRowIndex,
+                                        setProps,
+                                        mergeDuplicateHeaders,
+                                        column_selectable === 'single',
+                                        !allSelected
+                                    )}
+                                    name='column-header--select'
+                                    type={column_selectable === 'single' ?
+                                        'radio' :
+                                        'checkbox'
+                                    }
+                                />
+                            </span>
+                        }
+                        {sort_action === TableAction.None || !isLastRow ?
+                            null :
                             (<span
-                                className='sort'
+                                className='column-header--sort'
                                 onClick={doSort(column.id, sortBy, mode, setProps)}
                             >
                                 <FontAwesomeIcon icon={getSortingIcon(column.id, sortBy)} />
-                            </span>) :
-                            ''
+                            </span>)
                         }
 
-                        {renamable ?
+                        {!renamable ?
+                            null :
                             (<span
                                 className='column-header--edit'
-                                onClick={editColumnName(column, columns, headerRowIndex, setProps, mergeDuplicateHeaders)}
+                                onClick={editColumnName(column, visibleColumns, headerRowIndex, setProps, mergeDuplicateHeaders)}
                             >
                                 <FontAwesomeIcon icon='pencil-alt' />
-                            </span>) :
-                            ''
+                            </span>)
                         }
 
-                        {clearable ?
+                        {!clearable ?
+                            null :
                             (<span
                                 className='column-header--clear'
-                                onClick={doAction(actions.clearColumn, column, columns, headerRowIndex, mergeDuplicateHeaders, setFilter, setProps, map, data)}
+                                onClick={doAction(actions.clearColumn, selected_columns, column, visibleColumns, headerRowIndex, mergeDuplicateHeaders, setFilter, setProps, map, data)}
                             >
                                 <FontAwesomeIcon icon='eraser' />
-                            </span>) :
-                            ''
+                            </span>)
                         }
 
-                        {deletable ?
+                        {!deletable ?
+                            null :
                             (<span
                                 className={'column-header--delete' + (spansAllColumns ? ' disabled' : '')}
                                 onClick={spansAllColumns ?
                                     undefined :
-                                    doAction(actions.deleteColumn, column, columns, headerRowIndex, mergeDuplicateHeaders, setFilter, setProps, map, data)
+                                    doAction(actions.deleteColumn, selected_columns, column, visibleColumns, headerRowIndex, mergeDuplicateHeaders, setFilter, setProps, map, data)
                                 }
                             >
                                 <FontAwesomeIcon icon={['far', 'trash-alt']} />
-                            </span>) :
-                            ''
+                            </span>)
                         }
 
-                        {hideable ?
+                        {!hideable ?
+                            null :
                             (<span
                                 className={'column-header--hide' + (spansAllColumns ? ' disabled' : '')}
                                 onClick={spansAllColumns ?
                                     undefined :
                                     () => {
-                                        const ids = actions.getColumnIds(column, columns, headerRowIndex, mergeDuplicateHeaders);
+                                        const ids = actions.getColumnIds(column, visibleColumns, headerRowIndex, mergeDuplicateHeaders);
 
                                         const hidden_columns = hiddenColumns ?
                                             R.union(hiddenColumns, ids) :
@@ -207,8 +295,7 @@ function getter(
                                         setProps({ hidden_columns });
                                     }}>
                                 <FontAwesomeIcon icon={['far', 'eye-slash']} />
-                            </span>) :
-                            ''
+                            </span>)
                         }
 
                         <span className='column-header-name'>{labels[columnIndex]}</span>