+++ packages/dash-table/src/dash-table/derived/header/fixedHeaders.ts
@@ -1,9 +1,9 @@
-import { RowSelection } from 'dash-table/components/Table/props';
+import { Selection } from 'dash-table/components/Table/props';
 
 export default (
     fixedColumns: number,
     rowDeletable: boolean,
-    rowSelectable: RowSelection
+    rowSelectable: Selection
 ): number => {
     const offset =
         (rowDeletable ? 1 : 0) +