+++ packages/dash-table/tests/cypress/tests/server/delete_row_test.ts
@@ -10,7 +10,7 @@ describe('delete', () => {
     });
 
     it('can delete row when sorted', () => {
-        cy.get('tr th.column-0 .sort').last().click({ force: true }).click({ force: true });
+        cy.get('tr th.column-0 .column-header--sort').last().click({ force: true }).click({ force: true });
         DashTable.getCell(0, 0).within(() => cy.get('.dash-cell-value').should('have.html', '28155'));
         DashTable.getDelete(0).click();
         DashTable.getCell(0, 0).within(() => cy.get('.dash-cell-value').should('have.html', '28154'));